Let it compile on SPARC again Betarun version number: 48
git-svn-id: https://svn.nfit.au.dk/svn/beta/trunk@12877 678b6752-a8a0-49ff-a511-7c2a3241fcd4

+++ betarun/C/outpattern.c
@@ -1081,7 +1081,7 @@ int DisplayBetaStack(BetaErr errorNumber,
     }
   }
 
-  error_pc = (unsigned long)thePC;
+  error_pc = thePC;
 #ifdef RISC
   /* Correct PC in case of unalignment */
   thePC = (long *)((long)thePC & ~3);
